+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Gitaly changelog
 
+UNRELEASED
+
+- Implement CommitService.CommitLanguages, add gitaly-ruby
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/210
+
 v0.25.0
 
 - Implement FindAllTags RPC